src/portent/requirements.txt
discord.py==2.6.3
python-dotenv==1.1.1
rich==14.1.0
